feat: Refactor API paths to remove app prefix for consistency
